refactor: clean up, renaming
